diff

+++ predictor.cpp
@@ -27,10 +27,10 @@ void Predictor::printResults()
         << "Total predictions: " << d_totalPredictions << std::endl
         << "Number of mispredictions: " << d_mispredictions << std::endl
         << "Misprediction rate: " << std::fixed << std::setprecision(2)
-        << 100 * (static_cast<double>(d_mispredictions) / d_totalPredictions)
+        << 100 * (static_cast<double>(d_mispredictions) / static_cast<double>(d_totalPredictions))
         << "%" << std::endl
         << "Percentage of buckets_used: "
-        << 100 * (static_cast<double>(d_buckets_used.size()) / d_size) << "%"
+        << 100 * (static_cast<double>(d_buckets_used.size()) / static_cast<double>(d_size)) << "%"
         << std::endl;
 }
